% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_roughness.R
\name{optim_roughness}
\alias{optim_roughness}
\title{Compute the optimal smoothing parameter of slope_FUN}
\usage{
optim_roughness(y, slope_FUN, par = 0.5, avg_FUN = mean,
  method = c("Nelder-Mead"))
}
\arguments{
\item{y}{matrix of the signals.}

\item{slope_FUN}{function that compute the slope}

\item{par}{initial value of the smoothing paramter. Default is 0.5.}

\item{avg_FUN}{function to aggregate the rougness of each signal. Defaults is the methd.}

\item{method}{optimizer from \code{optim}. Defaults is "Nelder-Mead".}
}
\description{
Find smoothing paramters that matches rougness of original signal and of their slope
}
