% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.slopelm.R
\name{plot.slopelm}
\alias{plot.slopelm}
\title{Plot method for slopelm object}
\usage{
\method{plot}{slopelm}(x, effect = "all", type = "statistic",
  multcomp = NULL, alternative = "two.sided", enhanced_stat = FALSE,
  ...)
}
\arguments{
\item{x}{a slopelm object}

\item{effect}{a character string indicating the object to plot. Default is \code{"all"}.}

\item{type}{a character string indicating if the statistic of the coefficient should be plotted. Default is \code{"statistic"}.}

\item{multcomp}{a character string indicating which multiple comparison procedure should be plotted. Default is \code{"slope_clustermass"}.}

\item{alternative}{a character string indicating the alternative of the test in the case of a \code{t} test. Default is \code{"two.sided"}.}

\item{enhanced_stat}{a logical indicating if the enhanced statistic should be plotted instead of the univarite one. Default is \code{FALSE}.}

\item{...}{other argument pass to \code{plot}}
}
\description{
shows the statistics for the test on the mean and on the slope
}
