% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness.R
\name{roughness}
\alias{roughness}
\title{Roughness of Signals}
\usage{
roughness(x)
}
\arguments{
\item{x}{matrix of signals. The time is in column and each row represents one signal.}
}
\description{
Roughness of signals using the variance of the second derivative
}
