% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_diff}
\alias{slope_diff}
\title{Compute the slope of multiple signal using time difference}
\usage{
slope_diff(y)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}
}
\description{
compute the slope of multiple signal using the \code{diff} function.
}
