% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_locpol}
\alias{slope_locpol}
\title{Compute the slope of multiple signal using local polynormal}
\usage{
slope_locpol(y, bw = 0.01, kernel = EpaK, deg = 1)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}

\item{bw}{a scalar representing the bandwidth. The Default value is \code{.01}.}

\item{kernel}{a function representing for the kernel. By default the Epanechnikov Kernel.}

\item{deg}{a integer representing the degree of the polynomial use to fit the signal. Default value is \code{1}.}
}
\description{
compute the slope of multiple signal using local polynomial from the \code{locpol} package.
}
