% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_lpepa}
\alias{slope_lpepa}
\title{Compute the slope of multiple signal using local polynormal with Epanechnikov weights}
\usage{
slope_lpepa(y, bw = 0.01, deg = 1)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}

\item{bw}{a scalar representing the bandwidth as percentage of the signal. The Default value is \code{.01}.}

\item{deg}{a integer representing the degree of the polynomial use to fit the signal. Default value is \code{1}.}
}
\description{
compute the slope of multiple signal using local polynomial from the \code{lpepa} package.
}
