% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_roughtness}
\alias{slope_roughtness}
\title{Compute the slope of multiple signal using and adjust the smoothing paramter}
\usage{
slope_roughtness(y, slope_FUN = NULL)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}

\item{slope_FUN}{the function to compute the slope. The second argument of this function is the parameter to adjust.}
}
\description{
compute the slope of multiple signal and adjust the smoothing parameter to match the roughness from the original signal to the slope.
}
