% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_spectral}
\alias{slope_spectral}
\title{Compute the slopes of multiple signals using Fourrier Transform}
\usage{
slope_spectral(y)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}
}
\description{
Compute the slopes of multiple signals using the differentiation in the frequency space
}
