% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_spline}
\alias{slope_spline}
\title{Compute the slope of multiple signal using splines}
\usage{
slope_spline(y, spar = 0.5)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}

\item{spar}{a scalar representing the smoothing parameter. The Default value is \code{.5}.}
}
\description{
compute the slope of multiple signal using smothing spline
}
