% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopelm.R
\name{slopelm}
\alias{slopelm}
\title{Compute cluster-mass statistics using the mean and the slope of the signal}
\usage{
slopelm(
  formula,
  data = NULL,
  np = 5000,
  method = NULL,
  type = "permutation",
  test = "fisher",
  threshold = NULL,
  aggr_FUN = NULL,
  slope_FUN = slope_lpepa,
  multcomp = "slope",
  ...
)
}
\arguments{
\item{formula}{A formula object where the left part is a matrix defined in the global environnement.}

\item{data}{A data frame for the independant variables.}

\item{np}{The number of permutations. Default value is \code{5000}.}

\item{method}{A character string indicating the method used to handle nuisance variables. Default is \code{NULL} and will switch to \code{"freedman_lane"} for the fixed effects model and to \code{"Rd_kheradPajouh_renaud"} for the repeated measures ANOVA. See \link{lmperm} or \link{aovperm} for details on the permutation methods.}

\item{type}{A character string indicating the type of resampling. Default is \code{"permutation"} and \code{"signflip"} is also available.}

\item{test}{A character string to specify the name of the test. Default is \code{"fisher"}. \code{"t"} is available for the fixed effects model.}

\item{threshold}{A numerical vector that specify the limit of a cluster for the \code{"clustermass"} multiple comparisons procedure. If it is a vector each value will be associated to an effect. If it is vector the same threshold will be used for each test. Default value is \code{NULL} and will compute a threshold based on the 0.95 quantile of the choosen test statistic.}

\item{aggr_FUN}{A function that will be used to aggregate the statistics of a cluster into one scalar. Default is the sum of squares fot t statistic and sum for F statistic.}

\item{slope_FUN}{A function that will be used to compute the slope of the response variables. Default is  \code{slope_lpepa} (local polynomial) and \code{slope_spline} or \code{slope_diff} are also available.}

\item{multcomp}{A vector of character defining the methods of multiple comparisons to compute. Default is \code{"slope"}, and \code{"glue"} is an additional option available. \code{"slope"} considers tests on signals and slopes equally and \code{"glue"} uses tests on slopes only to bind clusters.}

\item{...}{Futher arguments, see details.}
}
\value{
A list containing : a table of the clusters, or a \code{multcomp} object for the other multiple comparison procedures. Use the \link{plot.slopelm} method to have a quick overview of the results.
}
\description{
Compute the clustermass statistic on the combined hypothesis that the slope or the mean of the signal are different from zero.
}
\details{
See the \link{clusterlm} function for detail about
}
\author{
jaromil.frossard@unige.ch
}
