% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_FUN_default.R
\name{slope_spectral}
\alias{slope_spectral}
\title{Compute the slope of multiple signal using Fourier}
\usage{
slope_spectral(y, ratio_max_mod = 10)
}
\arguments{
\item{y}{a matrix with multiple signals in rows and time in column}

\item{ratio_max_mod}{a scalar indicating the limit a frequency deleted based on the maximal value of the amplitude of the signal.}
}
\description{
compute the slope of multiple signal using the differentiation in the spectral space
}
